+++ Src/Library/changelog.md
@@ -1,3 +1,6 @@
+### NEW
+- `[Throttle(...)]` attribute for configuring endpoint #227
+
 ### FIX
 - pre/post processor collection modification bug #224
 - response dto initialization not working with array types #225